Modified the BOINC_GETSOCKOPT_TYPE macro so it determines the proper type of parameter 5 to getsockopt() by using the compiler.  The macro was in danger of becoming an ever expanding list of case statements, since some platforms use socklen_t, others use size_t, and still others use int.  Some aren't even consistent from OS rev to OS rev.
A macro BOINC_SOCKLEN_T is set in config.h to indicate the proper type.  This
is typedef to be boinc_socklen_t in lib/network.h.  The special cases for
__APPLE__ and WIN32 can probably be removed at a later time.  Places where
socklen_t was used have been changed to boinc_socklen_t.

Also added double inclusion protection to network.h


git-svn-id: http://boinc.berkeley.edu/svn/trunk/boinc@6055 677681c0-c3c0-48b4-bd46-92119ef6d97a