David this is a bug fix for the 'resending lost results' part of the scheduler.
Problem is that reply.wreq.core_client_version was only being set in
send_work(), which was too late for the resend_lost_work() part
of the code.  You might want to move all the initialization of reply.wreq
out of send_work().   The core client version is needed to see if the
app is compatible with it when calling get_app_version().


git-svn-id: http://boinc.berkeley.edu/svn/trunk/boinc@6867 677681c0-c3c0-48b4-bd46-92119ef6d97a